+++ scripts/lateralus.component.view.js
@@ -75,6 +75,8 @@ define([
      * @final
      */
     this.component = component;
+
+    this.delegateLateralusEvents();
     Backbone.View.call(this, options);
   };