+++ scripts/lateralus.mixins.js
@@ -165,5 +165,70 @@ define([
     }
   };
 
+  var delegateEventSplitter = /^(\S+)\s*(.*)$/;
+
+  /**
+   * Bind `{{#crossLink "Lateralus.mixins:lateralusEvents"}}{{/crossLink}}`, if
+   * it is defined.
+   * @method delegateLateralusEvents
+   * @chainable
+   * @private
+   */
+  mixins.delegateLateralusEvents = function () {
+    /**
+     * A map of functions or string references to functions that will handle
+     * [events](http://backbonejs.org/#Events) dispatched to the central
+     * `{{#crossLink "Lateralus"}}{{/crossLink}}` instance.     *
+     *     var ExtendedComponent = Lateralus.Component.extend({
+     *       name: 'extended'
+     *
+     *       ,lateralusEvents: {
+     *         anotherComponentChanged: 'onAnotherComponentChanged'
+     *
+     *         ,anotherComponentDestroyed: function () {
+     *           // ...
+     *         }
+     *       }
+     *
+     *       ,onAnotherComponentChanged: function () {
+     *         // ...
+     *       }
+     *     });
+     * @protected
+     * @property lateralusEvents
+     * @type {Object|undefined}
+     * @default undefined
+     */
+    var lateralusEvents = this.lateralusEvents;
+
+    if (!lateralusEvents) {
+      return;
+    }
+
+    for (var key in lateralusEvents) {
+      var method = lateralusEvents[key];
+      if (!_.isFunction(method)) {
+        method = this[lateralusEvents[key]];
+      }
+
+      if (!method) {
+        new Error('Method "' + method + '" not found for ' + this.toString());
+      }
+
+      var match = key.match(delegateEventSplitter);
+      var eventName = match[1];
+      var boundMethod = _.bind(method, this);
+
+      if (isLateralus(this)) {
+        this.on(eventName, boundMethod);
+      } else {
+        this.listenTo(this.lateralus, eventName, boundMethod);
+      }
+
+    }
+
+    return this;
+  };
+
   return mixins;
 });